refactor: Rename and update services page components
- Renamed DepannageServicesPage to ServicesPage for clarity.
- Updated import path for HeroSection to reflect new structure.
- Revised titles and descriptions in HeroSection for improved messaging and professionalism.
