+++ frontend/src/app/services/page.tsx
@@ -1,8 +1,8 @@
-import HeroSection from "@/components/page/services/depannage/HeroSection";
+import HeroSection from "@/components/page/services/HeroSection";
 import DetailsSection from "@/components/page/services/depannage/DetailsSection";
 import CTASection from "@/components/page/services/depannage/CTASection";
 
-export default function DepannageServicesPage() {
+export default function ServicesPage() {
   return (
     <main className="min-h-screen bg-background">
       <HeroSection />
